// src/lib/api/context/lib/types.ts
export type UUID = string;
// ...
export interface ContextClaims {
  tenant_id: UUID;
  workspace_id: UUID;
  sub: UUID; // user ID
}